+++ src/components/offers/offers.jsx
@@ -1,18 +1,13 @@
 import React from "react";
 import PropTypes from "prop-types";
 import Offer from "../../components/offer/offer.jsx";
-import withCitiesOffer from "../../hocs/with-cities-offer/with-cities-offer.js";
-import withCbOffer from "../../hocs/with-cb-offer/with-cb-offer.js";
-
-const OfferWrapped = withCbOffer(withCitiesOffer(Offer));
 
 const Offers = (props) => {
-  console.log(props.active)
   const {cityOffers, onSelect, onUnselect, onBookmarkClick} = props;
   return <div className="cities__places-list places__list tabs__content">
     {cityOffers.map((item, index) => {
       return (
-        <OfferWrapped
+        <Offer
           {...item}
           key={item[`id`]}
           onOfferHover={() => onSelect(index)}
